Reorganize to keep all construction code close together
