+++ emergence_lib/src/lib.rs
@@ -9,6 +9,7 @@
 #![allow(clippy::too_many_arguments)]
 
 pub mod asset_management;
+pub mod construction;
 pub mod crafting;
 pub mod curves;
 pub mod enum_iter;
