+++ emergence_lib/src/player_interaction/clipboard.rs
@@ -6,15 +6,13 @@ use leafwing_input_manager::prelude::ActionState;
 
 use crate::{
     asset_management::manifest::Id,
+    construction::{ghosts::Preview, terraform::TerraformingChoice},
     crafting::components::ActiveRecipe,
     simulation::geometry::{Facing, MapGeometry, TilePos},
-    structures::{construction::Preview, structure_manifest::Structure},
+    structures::structure_manifest::Structure,
 };
 
-use super::{
-    cursor::CursorPos, selection::CurrentSelection, terraform::TerraformingChoice,
-    InteractionSystem, PlayerAction,
-};
+use super::{cursor::CursorPos, selection::CurrentSelection, InteractionSystem, PlayerAction};
 
 /// Code and data for working with the clipboard
 pub(super) struct ClipboardPlugin;
@@ -111,7 +109,7 @@ impl Clipboard {
     /// Apply a tile-position shift to the items on the clipboard.
     ///
     /// Used to place items in the correct location relative to the cursor.
-    pub(super) fn offset_positions(&self, origin: TilePos) -> Vec<(TilePos, ClipboardData)> {
+    pub(crate) fn offset_positions(&self, origin: TilePos) -> Vec<(TilePos, ClipboardData)> {
         if let Clipboard::Structures(map) = self {
             map.iter()
                 .map(|(k, v)| ((*k + origin), v.clone()))
