+++ emergence_lib/src/structures/mod.rs
@@ -3,8 +3,10 @@
 //! Typically, these will produce and transform resources (much like machines in other factory builders),
 //! but they can also be used for defense, research, reproduction, storage and more exotic effects.
 
-use bevy::prelude::*;
+use bevy::{prelude::*, utils::HashSet};
 use bevy_mod_raycast::RaycastMesh;
+use hexx::{shapes::hexagon, Hex};
+use serde::{Deserialize, Serialize};
 
 use crate::{
     asset_management::{
@@ -12,20 +14,15 @@ use crate::{
         AssetCollectionExt,
     },
     player_interaction::{clipboard::ClipboardData, selection::ObjectInteraction},
-    simulation::{
-        geometry::{Facing, TilePos},
-        SimulationSet,
-    },
+    simulation::geometry::{Facing, TilePos},
 };
 
 use self::{
-    construction::{ghost_lifecycle, ghost_signals, validate_ghosts},
     structure_assets::StructureHandles,
     structure_manifest::{RawStructureManifest, Structure},
 };
 
 pub(crate) mod commands;
-pub mod construction;
 mod structure_assets;
 pub mod structure_manifest;
 
@@ -35,16 +32,7 @@ pub(super) struct StructuresPlugin;
 impl Plugin for StructuresPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(ManifestPlugin::<RawStructureManifest>::new())
-            .add_asset_collection::<StructureHandles>()
-            .add_systems(
-                (
-                    validate_ghosts,
-                    ghost_signals.after(validate_ghosts),
-                    ghost_lifecycle.after(validate_ghosts),
-                )
-                    .in_set(SimulationSet)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            );
+            .add_asset_collection::<StructureHandles>();
     }
 }
 
@@ -89,5 +77,57 @@ impl StructureBundle {
                 ..Default::default()
             },
         }
+    }
+}
+
+/// The set of tiles taken up by a structure.
+///
+/// Structures are always "centered" on 0, 0, so these coordinates are relative to that.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Footprint {
+    /// The set of tiles is taken up by this structure.
+    pub(crate) set: HashSet<TilePos>,
+}
+
+impl Default for Footprint {
+    fn default() -> Self {
+        Self::single()
+    }
+}
+
+impl Footprint {
+    /// A footprint that occupies a single tile.
+    pub fn single() -> Self {
+        Self {
+            set: HashSet::from_iter(vec![TilePos::ZERO]),
+        }
+    }
+
+    /// A footprint that occupies a set of tiles in a solid hexagon.
+    pub fn hexagon(radius: u32) -> Self {
+        let mut set = HashSet::new();
+        for hex in hexagon(Hex::ZERO, radius) {
+            set.insert(TilePos { hex });
+        }
+
+        Footprint { set }
+    }
+
+    /// Computes the set of tiles that this footprint occupies in world space, when centered at `center`.
+    pub(crate) fn in_world_space(&self, center: TilePos) -> HashSet<TilePos> {
+        self.set
+            .iter()
+            .map(|&offset| center + offset)
+            .collect::<HashSet<_>>()
+    }
+
+    /// Rotates the footprint by the provided [`Facing`].
+    pub(crate) fn rotated(&self, facing: Facing) -> Self {
+        let mut set = HashSet::new();
+        for &tile_pos in self.set.iter() {
+            set.insert(tile_pos.rotated(facing));
+        }
+
+        Footprint { set }
     }
 }
